Alternating Bit Protocol
Alternating Bit (ABT) Protocol is a networking TCP protocol used to transmit packets from transmitter A and Receiver B. This protocol uses a timer to keep track of the packets sent and the ACKs received for them. It keeps track of the packets using a bit, which means that it sends packets one at a time and uses two values to monitor the ACKs.