GBN.c
The Go Back N protocol is a transport layer protocol used to implement TCP. The transmitter A in this protocol maintains a window of frames to be transmitted to the Receiver B and using a single timer it keeps track of the ACK packets it receives, sequentially. Which means that if it receives the ACKs in order then it keeps on shifting its window by one packet. For every ACK received it resets its timer. If the timer expires or A gets duplicate ACKs, then A transmits its whole window.

B doesn't have to maintain any buffer to store the packets as it receives all the packets in order. If a packet is received out of order then B sends the ACK for only last correctly received packet.