Selective Repeat Protocol
This protocol is characterized by the way it sends its acknowledgements. The transmitter sends window of frames and gets to resend only those frames for which it either doesn't get an ACK. Every transmitted frame has got its own timer. This protocol is better than Go Back N as it doesn't flood the network with packets which have been received by the receiver previously.